// client/app/types/enums.ts
export enum RelationshipStatus {
  SINGLE = "single",
  IN_RELATIONSHIP = "inRelationship",
  ENGAGED = "engaged",
  MARRIED = "married",
  COMPLICATED = "complicated",
  DIVORCED = "divorced",
  WIDOWED = "widowed",
}

export enum Education {
  HIGH_SCHOOL = "highSchool",
  ASSOCIATES = "associatesDegree",
  BACHELORS = "bachelorsDegree",
  MASTERS = "mastersDegree",
  DOCTORATE = "doctorate",
  OTHER = "other",
}

export enum ResponseStatus {
  OK,
  ERROR,
}

export enum Gender {
  FEMALE = "Woman",
  MALE = "Man",
}